Support ContentLength Prefix
In order to both receive and write a file to the object store, the Router needs to know how much content there is since it will not encounter an `io.EOF` when the remote socket is trying to read bytes from it.

An 8 byte big-endian prefix provides the length of the content to follow. Since the object store limit is ~24 GB, the ~16 GB content limit per upload should be fine.

Updated documentation to reflect the change in the File Upload Packet Format from

```
[ 1458 Byte Data ]
```

to

```
[ 8 B ContentLength | 1450 byte Data ]
```