Support Object Delete
This adds object delete. There's the caveat of reducing the application to running on a single node. This is temporary.

Each feature adds a new port or a bespoke message format. To that end, a GRPC branch of effort is at a low simmer to remove the encoding concerns from developers (me?).