Cutover tickets to GRPC
All ticket operations Read/Write/Delete requests are now handled with GRPC. This reduces the TCP ports from 4 to two. The router uses port 5001 to listen for Write and Delete while a Read server runs on 5004.

The next phase of this is consolidating the router to a single port (5001) and using GRPC as the transport from clients