Multi-node Support
Round-robin multi-node support. Since all writes are blocking the slowest node has a heavy influence. In testing with two NAND nodes, the performance was pretty terrible but the write and the read were correct.

A simple fix is to co-routine each request tho it creates a write bias to the fastest node.